+++ crates/rust-client/src/transaction/mod.rs
@@ -455,8 +455,6 @@ pub struct TransactionStoreUpdate {
     executed_transaction: ExecutedTransaction,
     /// Block number at which the transaction was submitted.
     submission_height: BlockNumber,
-    /// Updated account state after the [`AccountDelta`] has been applied.
-    updated_account: Account,
     /// Information about note changes after the transaction execution.
     note_updates: NoteUpdateTracker,
     /// New note tags to be tracked.
@@ -469,21 +467,18 @@ impl TransactionStoreUpdate {
     /// # Arguments
     /// - `executed_transaction`: The executed transaction details.
     /// - `submission_height`: The block number at which the transaction was submitted.
-    /// - `updated_account`: The updated account state after applying the transaction.
     /// - `note_updates`: The note updates that need to be applied to the store after the
     ///   transaction execution.
     /// - `new_tags`: New note tags that were need to be tracked because of created notes.
     pub fn new(
         executed_transaction: ExecutedTransaction,
         submission_height: BlockNumber,
-        updated_account: Account,
         note_updates: NoteUpdateTracker,
         new_tags: Vec<NoteTagRecord>,
     ) -> Self {
         Self {
             executed_transaction,
             submission_height,
-            updated_account,
             note_updates,
             new_tags,
         }
@@ -497,11 +492,6 @@ impl TransactionStoreUpdate {
     /// Returns the block number at which the transaction was submitted.
     pub fn submission_height(&self) -> BlockNumber {
         self.submission_height
-    }
-
-    /// Returns the updated account.
-    pub fn updated_account(&self) -> &Account {
-        &self.updated_account
     }
 
     /// Returns the note updates that need to be applied after the transaction execution.
@@ -698,24 +688,16 @@ where
         info!("Applying transaction to the local store...");
 
         let account_id = tx_result.executed_transaction().account_id();
-        let account_delta = tx_result.account_delta();
         let account_record = self.try_get_account(account_id).await?;
 
         if account_record.is_locked() {
             return Err(ClientError::AccountLocked(account_id));
         }
 
-        let mut account: Account = account_record.into();
-        account.apply_delta(account_delta)?;
-
-        if self
-            .store
-            .get_account_header_by_commitment(account.commitment())
-            .await?
-            .is_some()
-        {
+        let final_commitment = tx_result.executed_transaction().final_account().commitment();
+        if self.store.get_account_header_by_commitment(final_commitment).await?.is_some() {
             return Err(ClientError::StoreError(StoreError::AccountCommitmentAlreadyExists(
-                account.commitment(),
+                final_commitment,
             )));
         }
 
@@ -739,7 +721,6 @@ where
         let tx_update = TransactionStoreUpdate::new(
             tx_result.into(),
             submission_height,
-            account,
             note_updates,
             new_tags,
         );